App: Move helpers to .util.helper package
